fix(frontend): apply workaround for wtr timeout bug
Applies a workaround for a known bug in the experimental Angular Web Test Runner builder where the `testsStartTimeout` setting is ignored.

Re-adds `zone.js` and `zone.js/testing` to the polyfills for the test configuration. This prevents the tests from hanging and timing out, allowing the test suite to run successfully in CI.